Verificato profilo SPCoop con nuovi jar 3parti

+++ tools/generic_project/src/org/openspcoop2/generic_project/io/IOUtilities.java
@@ -53,7 +53,9 @@ public class IOUtilities {
 			try{
 				if(fout!=null)
 					fout.close();
-			}catch(Exception e){}
+			}catch(Exception e){
+				// close
+			}
 		}
 	}
 
@@ -91,9 +93,11 @@ public class IOUtilities {
 
 			String files[] = srcPath.list();
 
-			for(int i = 0; i < files.length; i++){
-					IOUtilities.copyDirectory(new File(srcPath, files[i]), 
-							new File(dstPath, files[i]));
+			if(files!=null) {
+				for(int i = 0; i < files.length; i++){
+						IOUtilities.copyDirectory(new File(srcPath, files[i]), 
+								new File(dstPath, files[i]));
+				}
 			}
 		}
 		else{
@@ -219,11 +223,13 @@ public class IOUtilities {
     public static boolean deleteDir(File dir) {
     	 if (dir.isDirectory()) {
              String[] children = dir.list();
-             for (int i=0; i<children.length; i++) {
-                 boolean success = IOUtilities.deleteDir(new File(dir, children[i]));
-                 if (!success) {
-                     return false;
-                 }
+             if(children!=null) {
+	             for (int i=0; i<children.length; i++) {
+	                 boolean success = IOUtilities.deleteDir(new File(dir, children[i]));
+	                 if (!success) {
+	                     return false;
+	                 }
+	             }
              }
          }
